feat: Add "Add Tank" button to AppBar
This change adds an "Add Tank" button to the AppBar in the
TanksListScreen. The button is represented by the 'add_rounded' icon and
has a fixed width of 15. The purpose of this button is to enable the
user to easily add tanks to the tank list. This implementation is a TODO
and requires the implementation of the add tank function.